backend/requirements.txt
Copyright 2017 Google Inc.

Licensed under the Apache License, Version 2.0 (the "License");
you may not use this file except in compliance with the License.
You may obtain a copy of the License at

     http://www.apache.org/licenses/LICENSE-2.0

Unless required by applicable law or agreed to in writing, software
distributed under the License is distributed on an "AS IS" BASIS,
WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
See the License for the specific language governing permissions and
limitations under the License.
Flask==0.12.1
pyjwt==1.4.2
flask-cors==3.0.2
google-auth==0.10.0
requests==2.13.0
requests-toolbelt==0.7.1
python-firebase-gae==1.0.1
pyfcm==1.3.1
